unpack `position` as three floats
I noticed that 70% of rendering time is spent in the Vec3 constructor.
By passing the position as three floats (x,y,z) the allocation of a Vec3 isn't needed.

This gave a 4X speed improvement. Passing the color as components didn't have a large effect, but that
may depend on the specific colorizer.