Make API resemble Array more
renames `makeCellGrid` to `initialize`, add `repeat` and optimize
`matrixIndices`, shuffle the docs around a little to resemble `Array`.

This is breaking (requires a major release), so maybe rename
`initialize` back, but I like the matching of the `Array` api.

Maybe also `mapWithIndex` -> `indexedMap`, where the index is a
`(row, column)` tuple.

And the unpacking of `color` in RenderWebGL seems like a good idea.
The Vec3 instruction takes 30% of calculation time.

I can put those changes in this PR if you want.