Move away from TypedSvg
typed svg is more maintainable, but for packages, where all the ugly
bits are hidden away, I prefer the raw speed of `elm/svg`.

In particular because typed-svg uses avh4/elm-color, and its conversion
to css strings is slow. I'll open a PR for that, but now here it's fast.